Move `QuestionNotebookButton` component
