+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/QuestionNotebookButton/index.ts
